refactor(file structure): extract folder feature-status-bar
